Lazy creation of RTEs for inheritance children
Currently, expand_inherited_tables called from subquery_planner opens
all inheritance child tables and adds them to the query's range
table and PlannerInfo in the form of AppendRelInfo's initially and
later RelOptInfo's.  For partitioning, that's pretty wasteful, because
it's possible to determine only the partitions that will need to be
scanned using partition pruning, which doesn't require opening the
partitions themselves.  However, the point at which the above
processing happens is too early to perform partition pruning.

This commit rearranges things within the planner so that inheritance
child relations can be added to the Query and PlannerInfo during
make_one_rel, just before set_base_rel_sizes is called.  Although the
late initialization approach only benefits the partitioning case, it
seems undesirable to do it only for partitioned tables, because it
means adding code to handle partitions specially at various places
within the planner.  So, *all* inheritance parent tables are expanded
in make_one_rel.  All unpruned partitions are added to Query in the
form of their RangeTblEntry's being added to the range table and to
PlannerInfo in the form of an AppendRelInfo and a RelOptInfo for each.

For partitioning, although we don't create a RangeTblEntry and
RelOptInfo for pruned partitions at make_one_rel time, partitionwise
join code relies on the fact that even though partitions may have
been pruned, they'd still own a RelOptInfo to handle the outer join
case where the pruned partition appears on the nullable side of join.
Partitionwise join code deals with that by allocating dummy
RelOptInfos for pruned partitions that are based mostly on their
parent's properties.

There are some regression test diffs:

 1. Caused by the fact that we no longer allocate a duplicate RT
    entry for a partitioned table in its role as child, as seen in
    the partition_aggregate.out test output.

 2. Those in postgres_fdw.out are caused by the fact that junk columns
    required for row marking are added to reltarget->exprs later than
    user columns, because the row marking junk columns arent added
    until after the inheritance is expanded which as of this commit
    is later than it used to be.