Allow batching of inserts during cross-partition updates
Currently, the insert that runs internally as part of
cross-partition update of partitioned tables can't use batching
because ExecInitRoutingInfo(), which initializes the insert target
partition, allows batching only if the query's original operation is
CMD_INSERT.  This commit loosens that restriction to allow the
internal inserts of cross-partition updates use batching too,
essentially reverting a check that 927f453a94106 put in place.

While at it, also update the comment in
postgresGetForeignModifyBatchSize() a bit to clarify which inserts
are disallowed during cross-partition updates.

This also adjusts the test case added by 927f453a94106 to confirm
that the insert performed when moving the rows into the foreign
partition indeed uses batched mode.