Tweak handling of serialization failures during cascaded RI update/delete
The current coding for signaling a serializability failure during
cascaded update or delete of a row not visible to the start-of-the-
transaction snapshot involves a hack whereby the table AM forcefully
returns a status code that basically tells the executor that the row
has been concurrently updated such that the executor can issue an
error if needed.  However, the blocks of code that generate other
information to return about the failing row cannot do so in all cases
of the aforementioned hack, which can manifest as Assert failures in
those blocks.

To fix, simply error out immediately in the table AM itself rather
than in the caller, which avoids the gymnastics of returning the
correct failure information.

This also adds an isolation tests for some affected cases.