Handle tableoid=NULL for unmatched tuples in partitioned tables
For U/I/D, obviously all tuples have non-NULL tableoid, since that is
set to the OID of the partition that the tuple comes from.  However, for
MERGE that doesn't hold: all tuples in the source table that are
unmatched in the target table don't have a partition, so tableoid will
be NULL.  Cover that case too, by calling ExecMerge separately.

This is a kludge that will need to be handled differently.