Fix some issues around MERGE partition tuple routing
ExecMergeNotMatched() must perform the projection to emit the
"new" tuple to be inserted even before performing tuple routing,
because tuple routing expects to be passed a tuple matching the
target relation, in this case the root partitioned table.

While at it, instead of using ExecFindPartition() directly, rely
on the ExecInsert()'s ability to do the routing by itself.  That
also takes care of the cases where a routed tuple must be
converted into the destination leaf partition format before actually
inserting.