Revise MERGE initialization logic in ExecInitPartitionInfo()
This revises the entire MERGE initialization block in
ExecInitPartiitonInfo() to always create MergeActionState and
RelMergeActionState from scratch for a given partition being
initialized, instead of sometimes simply pointing to the first
result relation's instance and making own at other times.

Also, this fixes a thinko about which map to use when translating
the MergeAction expressions.  Currently, it uses the child-to-root-
relation map, though the source MergeAction item belongs to the
"first" leaf result relation, not the "root" relation.