Fix EXPLAIN failure in tests
A mysterious failure to show a plan turns out to be due to ruleutils.c
not being aware what the targetlist for MERGE is.  Inform it of that.

This is a bit ugly; perhaps there's a more reasonable answer.  But it
suffices for now, until further test cases inform us better.