Fix get_row_security_poilicies() for MERGE
It appeared that the existing relied on the RTE with index
mergeTarget_relation being marked with the necessary SELECT
security quals, though that seems wrong.  Given that it's a join RTE,
that never happened and thus the MERGE target relation never gets any
SELECT security quals assigned.

To fix, treat MERGE just like an UPDATE/DELETE as far whether and
which SELECT policy quals to apply is concerned.