Remove mergeSourceTargetList.
Apparently, this was kept around in the Query and subsequently in
the ModifyTable path and plan nodes so that set_plan_refs() could
use it to mark the MERGE action targetlist entries that refer to
the MERGE "source relation" attributes as INNER_VARs.  Though,
the ModifyTable subplan's targetlist can be used for that purpose
just fine, because that subplan is what constitutes the MERGE
"source relation".

This also teaches subquery pull-up code to process the MERGE action
expressions, because they may very well refer to the output of the
subqueries being pulled up.