Fix concurrent MERGE when update is cross-partition
ExecCrossPartitionUpdate() doesn't appear to work correctly
when called as part of an MERGE update action.  The fix is to
let ExecDelete() that runs as part of a cross-partition update
do the necessary EvalPlanQual() check instead of relying on
the caller ExecMergeMatched()'s code to do so, because that
doesn't quite work.

ExecCrossPartitionUpdate() projects the "retry" tuple that must be
returned to ExecUpdate() using the MERGE action info, instead of
the regular update projection.