Change the way resultRelInfo slots are initialized
In particular, use table_slot_create, not ExecInitExtraTupleSlot.  This
ensures we use the TTS routines that are correct for the relation.
(Doesn't matter in a world where everything is heapam, but better make
things consistent).  Also, rely on ri_projectNewInfoValid.

(This might be a problem if, say, ExecInsert ends up initializing only
the new slot and leaves the old one unset for some reason.  But if we
see that happening, we better fix that too.)