Fix handling of per-rel targetlists again
Revert 3123697cf73b.  I realize now that this change was ill-conceived,
because it sets things up incorrectly for when the output tuple needs to
be re-projected after EvalPlanQual. (Namely: the new tuple was being
projected from attributes out of merge's "subplan" targetlist, not from
the tuple obtained from the target relation's scan).

A proper fix to that problem requires completing the fix Amit did in
a85b2a6f5308: each rel's MergeAction targetlist needs to have its vars
adjusted during grouping_planner.

This accidentally breaks again a case that was working, so it still
needs a little bit more research.  But this is a major step forward.