Revert "Fix concurrent MERGE when update is cross-partition"
This reverts commit 33e904fe8e6e468fa78d97f5966c591da396572e,
because it seems ill-conceived now that I've seen why the
non-cross-partition update case bothers to perform EvalPlanQual()
in ExecMergeMatched() instead of in ExecUpdate().  It is to check
which of the WHEN quals (if any) the latest version of the
concurrently updated tuple satisfies and perform the respective
action, which may not be the same as the original action whose
update ran into the concurrent update.

This reinstates the "wrong" output for the cross-partition-update
test case that the commit being reverted was trying to address,
though a subsequent commit will implement the more correct fix for
that.

(Note: I didn't actually use git revert for this.)