Fix MERGE's EvalPlanQual() handling
Currently,ExecMergeMatched() passes ri_newTupleSlot, the new tuple
after applying the update, as the tuple to be checked to
EvalPlanQual(), though it should really be the latest tuple in the
update chain as other sites such as ExecUpdate()/ExecDelete() do.

As a result, this also adjusts the outputs of the
cross-partition-update test cases to reflect the "correct" expected
output.