Add per-result-relation extraUpdatedCols to ModifyTable
In spirit of removing things from RangeTblEntry that are better
carried by Query or PlannedStmt directly.

Instead of the rewriter populating extraUpdatedCols bitmapset in the
target relation's RTE, it is now planner's responsibility to compute
the bitmapset and pass it to the executor using the ModifyTable node.
In the inherited UPDATE's case, a copy is added for each child result
relation adjusted to account for any attribute number differences.