Do not add the NEW entry to view rule action's range table
The OLD entry suffices as a placeholder for the view relation when
it is queried, such as for checking its permissions during a query's
execution, but the NEW entry has no role whatsoever, so stop adding
it.

With there now being fewer entries in the view query's range table,
this change affects how the deparsed queries involving views look,
especially in the cases where the output of deparsing depends on using
RT indexs (such as automatically generated RTE alias names in
postgres_fdw deparser).  To wit, some postgres_fdw regression tests
whose expected output changes due to this have been updated to match.