Do not add the NEW entry to view rule action's range table
The OLD entry suffices as a placeholder for the view relation when
it is queried, such as checking its permissions during a query's
execution, but the NEW entry has no role to play whatsoever.

Because there are now fewer entries in the view query's range table,
this change affects how the deparsed queries look, where the output
of deparsing (such as alias names) depends on using RT indexs and
the original query references a view.  To wit, some postgres_fdw
regression tests whose expected output changes as a result of this
have been are updated to match.