Remove UpdateRangeTableOfViewParse()
And with it, the OLD and the NEW placeholder RTEs that are stored in
a view's stored rule action query.

The OLD RTE is needed currently as a placeholder for carrying view
relation info when querying a given view, because the view relation's
RTE in the query's range table is transposed into a subquery RTE for
converting the view's query into a subselect of the user query.

The NEW RTE is currently not needed for anything.

The purpose served by the OLD RTE doesn't really require it to be
present in the stored view rule query's range table though.
ApplyRetrieveRule(), which would so far look up the OLD RTE from the
view query and repurpose it for carrying the view relation info, can
create the placeholder entries (RTE and RTEPermissionInfo) by itself,
which this commit teaches it to do.

This changes the format of the store rule for views, especially the
range table length, so a bunch of regression tests that show deparsed
view queries need to be adjusted to cope.