Set inFromCl to false in child table RTEs
This is to allow the executor be able to distinguish tables that are
directly mentioned in the query from those that get added to the
query during planning.  A subsequent commit will teach the executor
to lock only the tables of the latter kind when executing a cached
plan.

Discussion: https://postgr.es/m/CA+HiwqFGkMSge6TgC9KQzde0ohpAycLQuV7ooitEEpbKB0O_mg@mail.gmail.comk