Adjustments to allow ExecutorStart() to sometimes fail
Upon passing a plan tree from a CachedPlan to the executor, there's a
possibility that ExecutorStart() might return an incompletely set up
planstate tree. This can happen if the CachedPlan undergoes invalidation
during the ExecInitNode() initialization process. In such cases, the
execution should be reattempted using a fresh CachedPlan. Also, any
partially initialized EState must be cleaned up by invoking both
ExecutorEnd() and FreeExecutorState().

ExecutorStart() (and ExecutorStart_hook()) now return a Boolean telling
the caller if the plan initialization failed.

For the replan loop in that context, it makes more sense to have
ExecutorStart() either in the same scope or closer to where
GetCachedPlan() is invoked.  So this commit modifies the following
sites:

* The ExecutorStart() call in ExplainOnePlan() is moved into a new
  function ExplainQueryDesc() along with CreateQueryDesc().  Callers
  of ExplainOnePlan() should now call the new function first.

* The ExecutorStart() call in _SPI_pquery() is moved to its caller
  _SPI_execute_plan().

* The ExecutorStart() call in PortalRunMulti() is moved to
  PortalStart().  This requires a new List field in PortalData to
  store the QueryDescs created in PortalStart() and a new memory
  context for those.  One unintended consequence is that
  CommandCounterIncrement() between queries in the PORTAL_MULTI_QUERY
  case is now done in the loop in PortalStart() and not in
  PortalRunMulti().  That still works because the Snapshot registered
  in QueryDesc/EState is updated to account for the CCI().

This commit also adds a new flag to EState called es_canceled that
complements es_finished to denote the new scenario where
ExecutorStart() returns with a partially setup planstate tree.  Also,
to reset the AFTER trigger state that would have been set up in the
ExecutorStart(), this adds a new function AfterTriggerCancelQuery()
which is called from ExecutorEnd() (not ExecutorFinish()) when
es_canceled is true.

Note that this commit by itself doesn't make any functional change,
because ExecutorStart() currently always returns true.  The changes
to make it check if the CachedPlan has become invalid will be
added in the upcoming patches.