Preparations to allow executor to take locks in some cases
This does two things on the executor side:

 * Make the executor take locks on the child tables if the plan
   comes from a CachedPlan.  To determine that a given range table
   relation is a child table, ExecGetRangeTableRelation() now
   examines RangeTblEntry.inFromCl.  To that end, the planner now
   sets inFromCl to false in the child tables' RTEs.  Also, the
   to deal with the cases where an unlocked child table might have
   been concurrently dropped, ExecGetRangeTableRelation() is now
   made to use try_table_open() replacing the existing table_open().

 * Add checks at various points during the executor's initialization
   of the plan tree to determine whether the originating CachedPlan
   has become invalid as a result of taking locks on the relations
   referenced in the plan.  This includes addding the check after
   every call to ExecOpenScanRelation() / ExecGetRangeTableRelation()
   and to ExecInitNode(), including the recursive ones to initialize
   child nodes.

If a given ExecInit*() function or any function called by it detects
that the plan has become invalid, it should return immediately even
if the PlanState node it's building may only be partially valid.
That is crucial for two reasons depending on where the check is:

 * The checks following ExecOpenScanRelation() /
   ExecGetRangeTableRelation() may find that the relation being
   opened has been dropped concurrently or that the plan has become
   invalid.  In this case, some operations in the code that follows
   may no longer be safe to do.  For example, it might try to
   dereference a NULL pointer in the case where the relation was
   dropped.

 * For the checks following ExecInitNode(), the returned child
   PlanState node might be only partially invalid.  The code that
   follows may misbehave if it depends on inspecting the child
   PlanState.  Note that this commit adds the check following all
   calls of ExecInitNode() that exist in the code base, even at
   sites where there is no code that might misbehave today, because
   it might misbehave in the future.  It seems like a good idea to
   put the guards in place today rather than in the future when the
   need arises.

To pass the CachedPlan that the executor will use for these checks,
ExecutorStart() (and ExecutorStart_hook) now gets a new parameter
CachedPlan *cplan.

Changes on the side of the callers of ExecutorStart():

ExecutorStart() (and ExecutorStart_hook()) now return a Boolean
telling the caller if the plan initialization failed.  When it
returns false due to the CachedPlan becoming invalid, the execution
should be reattempted using a fresh CachedPlan.  Actually, a new
function TryExecutorStart() is added to use by the call sites that
supply the PlannedStmt to execute from a CachedPlan, which takes
care of cleaning up the partially initialized execution state
including planstate tree.

For the replan loop in that context, it makes more sense to have
ExecutorStart() either in the same scope or closer to where
GetCachedPlan() is invoked.  So this commit modifies the following
sites:

* ExplainOnePlan() now returns a Boolean to tell the caller that
  TryExecutorStart() failed, so the caller should retry with a new
  plan.

* The ExecutorStart() call in _SPI_pquery() is moved to its caller
  _SPI_execute_plan() and replaced by TryExecutorStart().

* The ExecutorStart() call in PortalRunMulti() is moved to
  PortalStart() and replaced by TryExecutorStart().  This requires a
  new List field in PortalData to store the QueryDescs created in
  PortalStart() and a new memory context for those.  One unintended
  consequence is that CommandCounterIncrement() between queries in
  the PORTAL_MULTI_QUERY case is now done in the loop in PortalStart()
  and not in PortalRunMulti().  That still works because the Snapshot
  registered in QueryDesc/EState is updated to account for the CCI().

This commit also adds a new flag to EState called es_canceled that
complements es_finished to denote the new scenario where
ExecutorStart() returns with a partially setup planstate tree.  Also,
to reset the AFTER trigger state that would have been set up in the
ExecutorStart(), this adds a new function AfterTriggerCancelQuery()
which is called from ExecutorEnd() (not ExecutorFinish()) when
es_canceled is true.

Note that this commit by itself doesn't make any functional change,
because ExecutorStart() currently always returns true.  The changes
to make it check if the CachedPlan has become invalid will be
added in the upcoming patches.

Reviewed-by: Robert Haas

Discussion: https://postgr.es/m/CA+HiwqFGkMSge6TgC9KQzde0ohpAycLQuV7ooitEEpbKB0O_mg@mail.gmail.comk